跟新文档

+++ refresh-layout/src/main/java/com/scwang/smartrefresh/layout/util/DensityUtil.java
@@ -15,21 +15,27 @@ public class DensityUtil {
     }
 
     /**
-     * 根据手机的分辨率从 dp 的单位 转成为 px(像素) 
+     * 根据手机的分辨率从 dp 的单位 转成为 px(像素)
+     * @param dpValue 虚拟像素
+     * @return 像素
      */
     public static int dp2px(float dpValue) {
         return (int) (0.5f + dpValue * Resources.getSystem().getDisplayMetrics().density);
     }
 
     /**
-     * 根据手机的分辨率从 px(像素) 的单位 转成为 dp 
+     * 根据手机的分辨率从 px(像素) 的单位 转成为 dp
+     * @param pxValue 像素
+     * @return 虚拟像素
      */
-    public static float px2dp(float pxValue) {
+    public static float px2dp(int pxValue) {
         return (pxValue / Resources.getSystem().getDisplayMetrics().density);
     }
 
     /**
      * 根据手机的分辨率从 dp 的单位 转成为 px(像素)
+     * @param dpValue 虚拟像素
+     * @return 像素
      */
     public int dip2px(float dpValue) {
         return (int) (0.5f + dpValue * density);
@@ -37,8 +43,10 @@ public class DensityUtil {
 
     /**
      * 根据手机的分辨率从 px(像素) 的单位 转成为 dp
+     * @param pxValue 像素
+     * @return 虚拟像素
      */
-    public float px2dip(float pxValue) {
+    public float px2dip(int pxValue) {
         return (pxValue / density);
     }
 }
